+++ GHEtool/Validation/validate_deep_sizing.py
@@ -6,6 +6,7 @@ maximum peak temperature in injection and the average, undistrubed ground temper
 This can be understood since the main contributor to the peak temperature is the peak load, which, in the sizing,
 is expressed as a load per meter (W/m), so it scales like 1/borehole length.
 """
+import copy
 
 from GHEtool import *
 import matplotlib.pyplot as plt
@@ -31,23 +32,31 @@ def validate():
                                     col_injection=0, col_extraction=1)
     borefield.load = hourly_load
 
+    borefield_ct = copy.deepcopy(borefield)
+    borefield_ct.ground_data = GroundConstantTemperature(3, ground_data.calculate_Tg(20))
+
     # initiate lists
     Tg_list = []
-    max_Tf_list = []
+    Tg_list_ct = []
+    max_Tf_list_const = []
+    max_Tf_list_gradient = []
     length_list = range(20, 450, 20)
 
     for length in length_list:
         print(f'The current borehole length is {length} m.')
         borefield.calculate_temperatures(length)
+        borefield_ct.calculate_temperatures(length)
         Tg_list.append(
             borefield.ground_data.calculate_Tg(length + borefield.D, borefield.D))  # assume vertical boreholes
-        max_Tf_list.append(np.max(borefield.results.peak_injection))
+        Tg_list_ct.append(ground_data.calculate_Tg(20))
+        max_Tf_list_gradient.append(np.max(borefield.results.peak_injection))
+        max_Tf_list_const.append(np.max(borefield_ct.results.peak_injection))
 
     def f(x, a, b):
         return a / x + b
 
     # determine temperature difference between peak cooling temperature and ground temperature
-    diff = np.array(max_Tf_list) - np.array(Tg_list)
+    diff = np.array(max_Tf_list_gradient) - np.array(Tg_list)
 
     # fit to curve
     popt, pcov = curve_fit(f, length_list, diff)
@@ -55,10 +64,20 @@ def validate():
 
     plt.figure()
     plt.plot(length_list, Tg_list, label='Ground')
-    plt.plot(length_list, max_Tf_list, label='Fluid')
-    plt.hlines(borefield.Tf_max, 0, length_list[-1], label='Maximum temperature limit')
-    plt.xlabel('Borehole length [m]')
+    plt.plot(length_list, max_Tf_list_const, label='Fluid')
+    plt.plot(length_list, max_Tf_list_gradient, label='Combined effect')
+    plt.xlabel('Borehole depth [m]')
     plt.ylabel('Temperature [deg C]')
+    plt.title('Ground temperature gradient')
+    plt.legend()
+
+    plt.figure()
+    plt.plot(length_list, Tg_list_ct, label='Ground')
+    plt.plot(length_list, max_Tf_list_const, label='Fluid')
+    # plt.plot(length_list, max_Tf_list_const, label='Combined effect')
+    plt.xlabel('Borehole depth [m]')
+    plt.ylabel('Temperature [deg C]')
+    plt.title('Constant ground temperature')
     plt.legend()
     # plt.show()