+++ CHANGELOG.md
@@ -5,19 +5,37 @@ our [project board](https://github.com/users/wouterpeere/projects/2) on GitHub.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [2.3.3] - Unpublished
+## [2.3.4] - Unpublished
+
+### Added
+
+- Add flag in pressure drop calculation to in or exclude the pressure drop in the bend.
+
+### Fixed
+
+- Small bug in pressure drop lateral pipe.
+
+## [2.3.3] - 2025-05-28
 
 ### Added
 
 - Temperature dependent fluid properties (issue #143).
 - Freezing point to fluid data (issue #314).
+- Mass and volume percentage support(issue #350).
+- Add commercial fluids: Thermox DTX and Coolflow NTP from Hydratech.
 
 ### Changed
 
 - Changed back-end to be compatible with pygfunction 2.3.0 (issue #345).
 - The __repr__ of the different classes. It now returns a dictionary.
 - Remove ghe_logger since no longer used.
-- Add flag in pressure drop calculation to in or exclude the pressure drop in the bend.
+- Make Re a property in borefield just like Rb.
+- Calculate Re and Rb at minimum temperature or, if there is no temperature, the Tf_min.
+- Hydraulic calculations of Separatus.
+
+### Fixed
+
+- Fix export multiple U pipes (issue #352).
 
 ## [2.3.2] - 2025-04-02
 
@@ -29,6 +47,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Pressure drop calculation for horizontal pipe and total system (issue #332).
 - Added optimisation function for balanced borefield (issue #335).
 - Min_temperature and Max_temperature property to results class (issue #335).
+- Add PressureDrop class (issue #358).
 
 ### Changed
 
@@ -414,6 +433,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - fixed bug in interpolation
+
+[2.3.3]: https://github.com/wouterpeere/GHEtool/compare/v2.3.2...v2.3.3
 
 [2.3.2]: https://github.com/wouterpeere/GHEtool/compare/v2.3.1...v2.3.2